moved tooltipStyle into helpers/styling.ts; moved PPCOverlayTrigger.tsx into its own folder for consistency
